requirements.txt
pytest==8.3.2
python-dotenv==1.1.1
Requests==2.32.4
